Multiplication and power of 2 functions added

+++ pshapz/DTO/Sequence.cs
@@ -4,6 +4,6 @@
   {
     public Measure Literal { get; set; }
     public OperationSequence OperationType { get; set; }
-    public int ConstantValue { get; set; }
+    public decimal ConstantValue { get; set; }
   }
 }
